// launcher/src/types/nameSelect.types.ts
export interface NameSelectState {
  classId?: string;
  raceId?: string;
  raceName?: string;
  raceImage?: string;
  raceIcon?: string;
  deityId?: string;
  deityName?: string;
  deckId?: string;
  deckName?: string;
  characterName?: string;
}

export interface CharacterData {
  name: string;
  classId: string;
  className: string;
  raceId: string;
  raceName: string;
  raceImage: string;
  raceIcon: string;
  deityId: string;
  deityName: string;
  deckId: string;
  deckName: string;
}